backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coverage==7.8.0
cryptography==38.0.1
dj-database-url==1.0.0
Django==5.1.7
django-bootstrap5==23.3
django-livereload-server==0.5.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rosetta==0.9.8
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
Faker==15.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
polib==1.1.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.9.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
urllib3==1.26.12
whitenoise==6.2.0
